implement changes into og announcements page
