+++ src/frontend/src/routes.js
@@ -4,7 +4,6 @@ import LeagueManagementPage from "./views/LeagueManagementPage";
 import LoginPage from "./views/LoginPage";
 import MyTeamPage from "./views/MyTeamPage";
 import AnnouncementsPage from './views/AnnouncementsPage';
-import CommissionerAnnouncementsPage from './views/CommissionerAnnouncementsPage';
 import CreateAnnouncementsPage from './views/CreateAnnouncementsPage';
 import EditAnnouncementsPage from './views/EditAnnouncementsPage';
 
@@ -15,9 +14,8 @@ const routes = [
   { path: "/standings", component: StandingsPage },
   { path: "/manage", component: LeagueManagementPage, private: true },
   { path: '/announcements', component: AnnouncementsPage },
-  { path: '/manage/announcements', component: CommissionerAnnouncementsPage },// private: true },
-  { path: '/manage/announcements/edit/:id', component: EditAnnouncementsPage },// private: true },
-  { path: '/manage/announcements/create', component: CreateAnnouncementsPage },// private: true },
+  { path: '/announcements/edit/:id', component: EditAnnouncementsPage },// private: true },
+  { path: '/announcements/create', component: CreateAnnouncementsPage },// private: true },
 ];
 
 export default routes;
