+++ src/frontend/src/views/EditAnnouncementsPage.jsx
@@ -78,7 +78,7 @@ export default function EditAnnouncement() {
                 variant="outlined"
                 size="large"
                 sx={{ textTransform: 'none' }}
-                onClick={() => navigate('/manage/announcements')}
+                onClick={() => navigate('/announcements')}
               >
                 Cancel
               </Button>
